+++ src/Deriver/Runner.fs
@@ -5,7 +5,8 @@ module DeriverRunner =
 
     let runDeriver: RunPattern<DeriverApplication<'InputEvent, 'OutputEvent>, 'InputEvent, 'OutputEvent> =
         fun run (DeriverApplication application) ->
-            let beforeRun _ = ignore
+            let beforeRun _: BeforeRun<'InputEvent, 'OutputEvent> =
+                ignore
 
             application
             |> PatternRunner.runPattern (PatternName "Deriver") DeriverApplication.application beforeRun run