+++ src/Filter/Runner.fs
@@ -5,10 +5,11 @@ module FilterRunner =
 
     let runFilter: RunPattern<FilterApplication<'InputEvent, 'OutputEvent>, 'InputEvent, 'OutputEvent> =
         fun run (FilterApplication application) ->
-            let beforeRun filterApplication app =
-                filterApplication.FilterConfiguration
-                |> sprintf "%A"
-                |> app.Logger.VeryVerbose "Filter"
+            let beforeRun filterApplication: BeforeRun<'InputEvent, 'OutputEvent> =
+                fun app ->
+                    filterApplication.FilterConfiguration
+                    |> sprintf "%A"
+                    |> app.Logger.VeryVerbose "Filter"
 
             application
             |> PatternRunner.runPattern (PatternName "Filter") FilterApplication.application beforeRun run