+++ src/Router/Runner.fs
@@ -5,10 +5,11 @@ module ContentBasedRouterRunner =
 
     let runRouter: RunPattern<ContentBasedRouterApplication<'InputEvent, 'OutputEvent>, 'InputEvent, 'OutputEvent> =
         fun run (ContentBasedRouterApplication application) ->
-            let beforeRun routerApplication app =
-                routerApplication.RouterConfiguration
-                |> sprintf "%A"
-                |> app.Logger.VeryVerbose "Router"
+            let beforeRun routerApplication: BeforeRun<'InputEvent, 'OutputEvent> =
+                fun app ->
+                    routerApplication.RouterConfiguration
+                    |> sprintf "%A"
+                    |> app.Logger.VeryVerbose "Router"
 
             application
             |> PatternRunner.runPattern (PatternName "Router") ContentBasedRouterApplication.application beforeRun run